Add conversions for CompletionStage to Future which unwraps the CompletionException thrown by the CompletionStage.
In some cases failures contained in a CompletionStage are wrapped in a CompletionException. This makes conversions between Future and CompletionStage difficult as it changes the expected failure.

Side-note: The wrapping inside the CompletionStage does not always happen. it seems to happen only when adding an 'action' on the CompletionStage. If we convert a Future into a CompletionStage and back into a Future,without any in-between action, no CompletionException will be involved.

changelog_begin
changelog_end