Drop code to upload to com.digitalasset
We didn’t have any artifacts there since SDK 1.0 and we’re not
planning to add any. We suspect that the permission denied errors
might come from the fact that we need a different profile id for that
and while we don’t use the repo to upload anything we still create the
staging repo so this might fix that issue.

changelog_begin
changelog_end