kvutils/app: Create a test case for the Runner.
This is quite sophisticated because it needs to provision a
configuration, and so needs a basic implementation of the write->read
flow, at least for configuration updates.

The basic ledger used for the initial configuration update is
implemented using an `ArrayBuffer`, a `BoundedSourceQueue`, and a
`Dispatcher`, within the test file.

CHANGELOG_BEGIN
CHANGELOG_END