Remove usages of `Flowable.fromFuture` and `Single.fromFuture`
Fixes #12913

The resulting `Flowable`/`Single` is going to be blocking.

Since the `Future` is ultimately a `ListenableFuture`, we take advantage of the
possibility of adding a callback to it and we dispacth it to the execution
sequencer that is used to process response streams.

changelog_begin
[Java bindings] The Single resulting from calling several of the services
exposed by the bindings are no longer blocking. For further details, see
https://github.com/digital-asset/daml/issues/12193.
changelog_end