Make quickstart-java template work with canton sandbox. (#12453)
* Make quickstart-java template work with canton

Also update the quickstart-java release test instructions to use the new sandbox. This is a bit more involved than before, because of the party names being different every time, but I hope this is not too much of a burden. The test here is similar to what gets tested in the assistant integration test for quickstart-java.

changelog_begin
changelog_end

* try to fix indentation