Move more docs post processing to non-sphinx docs
This moves the redirects, the error page and the dummy versions file
to the non-sphinx html tarball which is handled in the split release
process.

This leaves only copying the pdf docs and the site map as a post
processing step which I cannot easily do here (since they should both
refer to the final artifact) so I’ll add that
directly in the assembly repo.

changelog_begin
changelog_end