Fix trigger compat test for canton sandbox
This fixes the timeout on trigger compat tests for the new sandbox (with LF >= 1.14).

The issue was that the trigger was "readAs"-ing with a party that
doesn't exist, since party ids aren't predictable with the new sandbox.

The solution is to allocate the Alice party ahead of time, and then
pass the allocated party in when calling `daml trigger`.

changelog_begin
changelog_end