split: do not tag
We use tags to track releases; in particular, we want tags to correspond
to GitHub releases. In the split release process, the `assembly` repo is
responsible for creating GitHub releases, so it would make sense that it
is also the one responsible for creating tags.

I wouldd also like to take this opportunity to manually remove all the
tags that have been created by the split release process so far and that
don't correspond to any release (which is all of them as we haven't done
a full split release yet), since they have already caused some
confusion.

CHANGELOG_BEGIN
CHANGELOG_END