Drop dummy token from JSON API check
That token is broken in a number of ways, e.g. broken ledger ids,
parties that may not exist, ….

When we created that there was no healthcheck endpoint on the JSON API
but now that there is one, we can just use that one and it doesn’t
need any token so things get a lot simpler.

changelog_begin
changelog_end