ledger-on-memory: Do not create iterators on a mutable log.
We must copy the slice of the log that we are reading, to ensure that we
are not iterating over the log while we mutate it in a separate
transaction. (In other words, objects must not escape the `readLog`
function.)

This is most likely the cause of a flaky test run that we saw recently.

CHANGELOG_BEGIN
CHANGELOG_END