Engine: fix compilation exception catch
During refacotring of #9993, the "unsafe" compilation of the command
as been moved out of the scope of `Engine.runSafely`, whose only purpose
is to catch exceptions thrown by the Speedy compiler.

This fixes the issue and rename the function `runSafely` to
`runCompilkerSafely` to make its purpose more obvious.

This is not a critical bug, as the command preprocessing should ensure
no error are thrown by the engine.

CHANGELOG_BEGIN
CHANGELOG_END