split release
This is an out-of-schedule releasee to get the breaking changes in
f2b7902c68f74fedee44cd262cc006e0c0a6765c (#12682) in a PR so Canton can
build on it and make a release we can depend on to f i x #12813.

Or something like that.

CHANGELOG_BEGIN
CHANGELOG_END