Speedy: Fix computation of Node version (#12879)
The computation of Node version relies on the
`packageToTransactionVersion` field of the PartialTransaction.  This
is initialized during the Speedy `Machine` construction with the
compiledPackage used to initialized the Machine. However the
compiledPackage is a mutable field and can be udpated during
interpreation, potentially making the 'PartialTransaction''s
`packageToTransactionVersion` out of date.

This is not a production bug, as production code use the Machine's
`CompiledPackage` in a imutable way (with a mutable
ConcurrentCompilePackage), however this is a problem for testing code
that may use the field in a mutable way (with an immutable
`PureCompiledPackage`).

We solve the issue by making the computation of the version, outise
the PartialTransaction object.

CHANGELOG_BEGING
CHANGELOG_END