Wire contract identifier autocompletion in Navigator
Navigator already had the possibility of autocompleting contract
identifiers, the only thing missing was to wire it in in full. This is
done mostly to inform how to address the more involved #12789.

changelog_begin
[Navigator] All input fields where a contract identifier can be passed
now provide the possibility of auto-completing based on the contracts
existing on the ledger.
changelog_end