No parties in create-daml-app's daml.yaml template,
... so that Navigator shows users for login by default.

When parties are specified explicitly in daml.yaml,
navigator uses those instead of querying user management
for the list of users a user can log in as.

These days, it actually doesn't make sense to list the
parties in the daml.yaml. Navigator will query party
mgmt for a list of known parties. We don't implicitly
allocate parties (anymore), so unclear why this section
is still here.

The users for the template are created by the init script
(see Setup.daml).

See #12945

CHANGELOG_BEGIN
CHANGELOG_END