Fix MacOS GHCi issues
The trick is extremely simple actually, see the comment for details.
Pretty embarassed that it took me this long to figure it out but
better late than never I guess.

I also upgraded grpc to match our java version and dropped the comment
to keep it in sync with the nix version which we hopefully never need
now. Not strictly required but seemed nice to do at the same time.

changelog_begin
changelog_end