Refactor Scala codegen: simplify by removing unnecessary abstractions (#13361)
* Refactor Scala codegen

changelog_begin
changelog_end

* Remove unnecessary abstractions

`DependencyGraph` and `LFUtil` both had an abstract superclass of
which they were the only implementation. Removing this unneccesary
layer of abstraction allows to avoid jumping between the superclass
and its concrete implementation to make sense of the overall meaning
of the code. Removing this layer also allows to find and remove
dead code and unused parameters.

* Throw away dead code around `BaseNode` and remove the unnecessary abstraction

* Throw away unused code

* Fix typo: unsopported -> unsupported

* Remove incorrect capitalization of type definitions

Before the refactoring, both `Util` and `LFUtil` had an implementation
of `toTypeDef` which were mostly matching, apart from the fact that
one capitalized its output while the other didn't. Apparently, I used
the wrong version and this broke existing code.