split out scalatest-utils from bazel build
This is one possible approach to taking things gradually out of the main
Bazel build.

This approach is focused on making the most-localized-possible change,
because that reduces the possible "blast radius" of what might be
considered a pretty risky move. In particular:

- The library is still published to Maven by the Bazel build, with the
  same version number as the rest of the SDK, ensuring that we don't
  break anything downstream.\*
- It is still very easy to develop the library alongside the rest of the
  repo, i.e. the default is for Bazel to rely on the local version of
  the library.
- Despite the previous point, we do not break Bazel caching for people
  who do not modify this library (including CI).

If we did choose this approach, this would only be the first step. The
next steps would be (roughly 1 PR per step, but preferably each step
would go through a full snapshot release cycle to ensure nothing gets
broken):

1. Start publishing this library from SBT, _in addition to_ the one
   Bazel publishes (with a slightly different version scheme), on the
   same schedule.
2. Change the Bazel build to stop publishing its own version of this
   library. Note that even as of this PR, the Bazel build does not
   depend on the artifact being on Maven, so this should not break the
   build (but may affect downstream consumers that have expectations about
   version strings across components).
3. Clean-up: there's a fair amount of work being one here to give Bazel
   access to everything it needs to publish to Maven, that could get
   simplified at this point.
4. Desynchronize publication: publish new versions of the library on a
   different schdule, dictated by the library itself rather than the
   repo as a whole.
5. Stop Bazel from depending on the state of local files, and instead
   depend on the latest version published to Maven Central.
6. Move the build to its own separate repo, as at this point there's no
   real link between the two builds anymore.

This is a very slow, careful approach, that assumes it is important to
retain the ability to work on the library simultaneously with the rest
of the repo for a while (up to step 5 in this case).

If we do not make that assumption and instead opt for a more "rip off
the bandaid" approach, this can be massively simplified. We could do:

1. Instead of this PR, simply remove the code for the library from this
   repo entirely, and fix the dependency on the latest published
   version (as there already is one).
2. Create a new repo that starts publishing new versions with its own CI
   and build system.

This would be slightly more risky for the overall release process and
downstream dependencies, as well as more disruptive to existing working
habits for this repo, but would require a lot less work and would allow
us to move a lot faster.

My personal preference would be to go for the faster option, but I did
want to present all options to open up the discussion.

\*: This is, of course, assuming I didn't screw up. I can't test the
actual release process without actually making a release.

CHANGELOG_BEGIN
CHANGELOG_END