Extract the participant test context interface.
In a follow up PR we will add an implementation which adds a timeout to each public API method from the participant test context.
The purpose is to reduce the number of builds that fail with generic test timeouts and instead fail fast when calls (especially streaming calls) take longer than expected.

CHANGELOG_BEGIN
CHANGELOG_END