remove scalatest-utils
This is the "fast" alternative to #13348. Though conceptually more
radical, this results in a much simpler set of changes. Importantly,
this does not touch the release process, meaning there is much less of a
chance that this would break the release process, particularly for patch
releases.

This is, of course, not the full story: the code removed here would need
to be added to some other repo.

CHANGELOG_BEGIN
CHANGELOG_END