Restrict Java codegen to types that can be seen over the Ledger API
Fixes #13324

Builds on top of refactorings part of #13372, #13439 and #13440,
alongside adding interface support to the transitive closure function
in #13557.

changelog_begin
[Java codegen] The Java codegen now restricts its output to Daml definitions
that are either directly or indirectly referred as part of a template or
interface (that is, only the types that can be seen over the Ledger API).
See #13324 for further details.
changelog_end